@@ -1,6 +1,6 @@
 module github.com/bwmarrin/discordgo/examples/linked_roles
 
-go 1.13
+go 1.24
 
 replace github.com/bwmarrin/discordgo v0.26.1 => ../../
 
