@@ -4,6 +4,7 @@ import (
 	"encoding/binary"
 	"encoding/json"
 	"fmt"
+	"github.com/nyttikord/gokord/logger"
 	"net"
 	"strconv"
 	"strings"
@@ -22,8 +23,7 @@ import (
 type VoiceConnection struct {
 	sync.RWMutex
 
-	Debug        bool // If true, print extra logging -- DEPRECATED
-	LogLevel     int
+	LogLevel     logger.Level
 	Ready        bool // If true, voice is ready to send/receive audio
 	UserID       string
 	GuildID      string
@@ -69,8 +69,7 @@ type VoiceSpeakingUpdateHandler func(vc *VoiceConnection, vs *VoiceSpeakingUpdat
 // once finished sending audio.
 // b : Send true if speaking, false if not.
 func (v *VoiceConnection) Speaking(b bool) (err error) {
-
-	v.log(LogDebug, "called (%t)", b)
+	v.LogDebug("called (%t)", b)
 
 	type voiceSpeakingData struct {
 		Speaking bool `json:"speaking"`
@@ -95,7 +94,7 @@ func (v *VoiceConnection) Speaking(b bool) (err error) {
 	defer v.Unlock()
 	if err != nil {
 		v.speaking = false
-		v.log(LogError, "Speaking() write json error, %s", err)
+		v.LogError("Speaking() write json error, %s", err)
 		return
 	}
 
@@ -107,8 +106,7 @@ func (v *VoiceConnection) Speaking(b bool) (err error) {
 // ChangeChannel sends Discord a request to change channels within a Guild
 // !!! NOTE !!! This function may be removed in favour of just using ChannelVoiceJoin
 func (v *VoiceConnection) ChangeChannel(channelID string, mute, deaf bool) (err error) {
-
-	v.log(LogInformational, "called")
+	v.LogDebug("called")
 
 	data := voiceChannelJoinOp{4, voiceChannelJoinData{&v.GuildID, &channelID, mute, deaf}}
 	v.session.wsMutex.Lock()
@@ -128,7 +126,6 @@ func (v *VoiceConnection) ChangeChannel(channelID string, mute, deaf bool) (err 
 // Disconnect disconnects from this voice channel and closes the websocket
 // and udp connections to Discord.
 func (v *VoiceConnection) Disconnect() (err error) {
-
 	// Send a OP4 with a nil channel to disconnect
 	v.Lock()
 	if v.sessionID != "" {
@@ -143,7 +140,7 @@ func (v *VoiceConnection) Disconnect() (err error) {
 	// Close websocket and udp connections
 	v.Close()
 
-	v.log(LogInformational, "Deleting VoiceConnection %s", v.GuildID)
+	v.LogInfo("Deleting VoiceConnection %s", v.GuildID)
 
 	v.session.Lock()
 	delete(v.session.VoiceConnections, v.GuildID)
@@ -154,8 +151,7 @@ func (v *VoiceConnection) Disconnect() (err error) {
 
 // Close closes the voice ws and udp connections
 func (v *VoiceConnection) Close() {
-
-	v.log(LogInformational, "called")
+	v.LogDebug("called")
 
 	v.Lock()
 	defer v.Unlock()
@@ -164,22 +160,22 @@ func (v *VoiceConnection) Close() {
 	v.speaking = false
 
 	if v.close != nil {
-		v.log(LogInformational, "closing v.close")
+		v.LogDebug("closing v.close")
 		close(v.close)
 		v.close = nil
 	}
 
 	if v.udpConn != nil {
-		v.log(LogInformational, "closing udp")
+		v.LogDebug("closing udp")
 		err := v.udpConn.Close()
 		if err != nil {
-			v.log(LogError, "error closing udp connection, %s", err)
+			v.LogWarn("error closing udp connection, %s", err)
 		}
 		v.udpConn = nil
 	}
 
 	if v.wsConn != nil {
-		v.log(LogInformational, "sending close frame")
+		v.LogDebug("sending close frame")
 
 		// To cleanly close a connection, a client should send a close
 		// frame and wait for the server to close the connection.
@@ -187,16 +183,16 @@ func (v *VoiceConnection) Close() {
 		err := v.wsConn.WriteMessage(websocket.CloseMessage, websocket.FormatCloseMessage(websocket.CloseNormalClosure, ""))
 		v.wsMutex.Unlock()
 		if err != nil {
-			v.log(LogError, "error closing websocket, %s", err)
+			v.LogError("error closing websocket, %s", err)
 		}
 
 		// TODO: Wait for Discord to actually close the connection.
 		time.Sleep(1 * time.Second)
 
-		v.log(LogInformational, "closing websocket")
+		v.LogDebug("closing websocket")
 		err = v.wsConn.Close()
 		if err != nil {
-			v.log(LogError, "error closing websocket, %s", err)
+			v.LogError("error closing websocket, %s", err)
 		}
 
 		v.wsConn = nil
@@ -242,8 +238,7 @@ type voiceOP2 struct {
 // WaitUntilConnected waits for the Voice Connection to
 // become ready, if it does not become ready it returns an err
 func (v *VoiceConnection) waitUntilConnected() error {
-
-	v.log(LogInformational, "called")
+	v.LogDebug("called")
 
 	i := 0
 	for {
@@ -267,15 +262,14 @@ func (v *VoiceConnection) waitUntilConnected() error {
 // after VoiceChannelJoin is used and the data VOICE websocket events
 // are captured.
 func (v *VoiceConnection) open() (err error) {
-
-	v.log(LogInformational, "called")
+	v.LogDebug("called")
 
 	v.Lock()
 	defer v.Unlock()
 
 	// Don't open a websocket if one is already open
 	if v.wsConn != nil {
-		v.log(LogWarning, "refusing to overwrite non-nil websocket")
+		v.LogWarn("refusing to overwrite non-nil websocket")
 		return
 	}
 
@@ -298,11 +292,11 @@ func (v *VoiceConnection) open() (err error) {
 
 	// Connect to VoiceConnection Websocket
 	vg := "wss://" + strings.TrimSuffix(v.endpoint, ":80")
-	v.log(LogInformational, "connecting to voice endpoint %s", vg)
+	v.LogDebug("connecting to voice endpoint %s", vg)
 	v.wsConn, _, err = v.session.Dialer.Dial(vg, nil)
 	if err != nil {
-		v.log(LogWarning, "error connecting to voice endpoint %s, %s", vg, err)
-		v.log(LogDebug, "voice struct: %#v\n", v)
+		v.LogError("error connecting to voice endpoint %s, %s", vg, err)
+		v.LogDebug("voice struct: %#v\n", v)
 		return
 	}
 
@@ -322,7 +316,7 @@ func (v *VoiceConnection) open() (err error) {
 	err = v.wsConn.WriteJSON(data)
 	v.wsMutex.Unlock()
 	if err != nil {
-		v.log(LogWarning, "error sending init packet, %s", err)
+		v.LogError("error sending init packet, %s", err)
 		return
 	}
 
@@ -339,8 +333,7 @@ func (v *VoiceConnection) open() (err error) {
 // wsListen listens on the voice websocket for messages and passes them
 // to the voice event handler.  This is automatically called by the Open func
 func (v *VoiceConnection) wsListen(wsConn *websocket.Conn, close <-chan struct{}) {
-
-	v.log(LogInformational, "called")
+	v.LogDebug("called")
 
 	for {
 		_, message, err := v.wsConn.ReadMessage()
@@ -348,7 +341,7 @@ func (v *VoiceConnection) wsListen(wsConn *websocket.Conn, close <-chan struct{}
 			// 4014 indicates a manual disconnection by someone in the guild;
 			// we shouldn't reconnect.
 			if websocket.IsCloseError(err, 4014) {
-				v.log(LogInformational, "received 4014 manual disconnection")
+				v.LogInfo("received 4014 manual disconnection")
 
 				// Abandon the voice WS connection
 				v.Lock()
@@ -356,7 +349,7 @@ func (v *VoiceConnection) wsListen(wsConn *websocket.Conn, close <-chan struct{}
 				v.Unlock()
 
 				// Wait for VOICE_SERVER_UPDATE.
-				// When the bot is moved by the user to another voice channel,
+				// When the user moves the bot to another voice channel,
 				// VOICE_SERVER_UPDATE is received after the code 4014.
 				for i := 0; i < 5; i++ { // TODO: temp, wait for VoiceServerUpdate.
 					<-time.After(1 * time.Second)
@@ -367,12 +360,12 @@ func (v *VoiceConnection) wsListen(wsConn *websocket.Conn, close <-chan struct{}
 					if !reconnected {
 						continue
 					}
-					v.log(LogInformational, "successfully reconnected after 4014 manual disconnection")
+					v.LogInfo("successfully reconnected after 4014 manual disconnection")
 					return
 				}
 
 				// When VOICE_SERVER_UPDATE is not received, disconnect as usual.
-				v.log(LogInformational, "disconnect due to 4014 manual disconnection")
+				v.LogInfo("disconnect due to 4014 manual disconnection")
 
 				v.session.Lock()
 				delete(v.session.VoiceConnections, v.GuildID)
@@ -384,16 +377,16 @@ func (v *VoiceConnection) wsListen(wsConn *websocket.Conn, close <-chan struct{}
 			}
 
 			// Detect if we have been closed manually. If a Close() has already
-			// happened, the websocket we are listening on will be different to the
+			// happened, the websocket we are listening to on will be different to the
 			// current session.
 			v.RLock()
 			sameConnection := v.wsConn == wsConn
 			v.RUnlock()
 			if sameConnection {
 
-				v.log(LogError, "voice endpoint %s websocket closed unexpectantly, %s", v.endpoint, err)
+				v.LogError("voice endpoint %s websocket closed unexpectantly, %s", v.endpoint, err)
 
-				// Start reconnect goroutine then exit.
+				// Start to reconnect goroutine then exit.
 				go v.reconnect()
 			}
 			return
@@ -412,21 +405,18 @@ func (v *VoiceConnection) wsListen(wsConn *websocket.Conn, close <-chan struct{}
 // wsEvent handles any voice websocket events. This is only called by the
 // wsListen() function.
 func (v *VoiceConnection) onEvent(message []byte) {
-
-	v.log(LogDebug, "received: %s", string(message))
+	v.LogDebug("received: %s", string(message))
 
 	var e Event
 	if err := json.Unmarshal(message, &e); err != nil {
-		v.log(LogError, "unmarshall error, %s", err)
+		v.LogError("unmarshall error, %s", err)
 		return
 	}
 
 	switch e.Operation {
-
 	case 2: // READY
-
 		if err := json.Unmarshal(e.RawData, &v.op2); err != nil {
-			v.log(LogError, "OP2 unmarshall error, %s, %s", err, string(e.RawData))
+			v.LogError("OP2 unmarshall error, %s, %s", err, string(e.RawData))
 			return
 		}
 
@@ -437,7 +427,7 @@ func (v *VoiceConnection) onEvent(message []byte) {
 		// Start the UDP connection
 		err := v.udpOpen()
 		if err != nil {
-			v.log(LogError, "error opening udp connection, %s", err)
+			v.LogError("error opening udp connection, %s", err)
 			return
 		}
 
@@ -456,7 +446,6 @@ func (v *VoiceConnection) onEvent(message []byte) {
 
 			go v.opusReceiver(v.udpConn, v.close, v.OpusRecv)
 		}
-
 		return
 
 	case 3: // HEARTBEAT response
@@ -469,7 +458,7 @@ func (v *VoiceConnection) onEvent(message []byte) {
 
 		v.op4 = voiceOP4{}
 		if err := json.Unmarshal(e.RawData, &v.op4); err != nil {
-			v.log(LogError, "OP4 unmarshall error, %s, %s", err, string(e.RawData))
+			v.LogError("OP4 unmarshall error, %s, %s", err, string(e.RawData))
 			return
 		}
 		return
@@ -481,7 +470,7 @@ func (v *VoiceConnection) onEvent(message []byte) {
 
 		voiceSpeakingUpdate := &VoiceSpeakingUpdate{}
 		if err := json.Unmarshal(e.RawData, voiceSpeakingUpdate); err != nil {
-			v.log(LogError, "OP5 unmarshall error, %s, %s", err, string(e.RawData))
+			v.LogError("OP5 unmarshall error, %s, %s", err, string(e.RawData))
 			return
 		}
 
@@ -490,7 +479,7 @@ func (v *VoiceConnection) onEvent(message []byte) {
 		}
 
 	default:
-		v.log(LogDebug, "unknown voice operation, %d, %s", e.Operation, string(e.RawData))
+		v.LogWarn("unknown voice operation, %d, %s", e.Operation, string(e.RawData))
 	}
 
 	return
@@ -508,7 +497,6 @@ type voiceHeartbeatOp struct {
 // is still connected.  If you do not send these heartbeats Discord will
 // disconnect the websocket connection after a few seconds.
 func (v *VoiceConnection) wsHeartbeat(wsConn *websocket.Conn, close <-chan struct{}, i time.Duration) {
-
 	if close == nil || wsConn == nil {
 		return
 	}
@@ -517,12 +505,12 @@ func (v *VoiceConnection) wsHeartbeat(wsConn *websocket.Conn, close <-chan struc
 	ticker := time.NewTicker(i * time.Millisecond)
 	defer ticker.Stop()
 	for {
-		v.log(LogDebug, "sending heartbeat packet")
+		v.LogDebug("sending heartbeat packet")
 		v.wsMutex.Lock()
 		err = wsConn.WriteJSON(voiceHeartbeatOp{3, int(time.Now().Unix())})
 		v.wsMutex.Unlock()
 		if err != nil {
-			v.log(LogError, "error sending heartbeat to voice endpoint %s, %s", v.endpoint, err)
+			v.LogError("error sending heartbeat to voice endpoint %s, %s", v.endpoint, err)
 			return
 		}
 
@@ -560,7 +548,6 @@ type voiceUDPOp struct {
 // and can be used to send or receive audio.  This should only be called
 // from voice.wsEvent OP2
 func (v *VoiceConnection) udpOpen() (err error) {
-
 	v.Lock()
 	defer v.Unlock()
 
@@ -583,14 +570,14 @@ func (v *VoiceConnection) udpOpen() (err error) {
 	host := v.op2.IP + ":" + strconv.Itoa(v.op2.Port)
 	addr, err := net.ResolveUDPAddr("udp", host)
 	if err != nil {
-		v.log(LogWarning, "error resolving udp host %s, %s", host, err)
+		v.LogError("error resolving udp host %s, %s", host, err)
 		return
 	}
 
-	v.log(LogInformational, "connecting to udp addr %s", addr.String())
+	v.LogDebug("connecting to udp addr %s", addr.String())
 	v.udpConn, err = net.DialUDP("udp", nil, addr)
 	if err != nil {
-		v.log(LogWarning, "error connecting to udp addr %s, %s", addr.String(), err)
+		v.LogError("error connecting to udp addr %s, %s", addr.String(), err)
 		return
 	}
 
@@ -603,7 +590,7 @@ func (v *VoiceConnection) udpOpen() (err error) {
 	// And send that data over the UDP connection to Discord.
 	_, err = v.udpConn.Write(sb)
 	if err != nil {
-		v.log(LogWarning, "udp write error to %s, %s", addr.String(), err)
+		v.LogError("udp write error to %s, %s", addr.String(), err)
 		return
 	}
 
@@ -614,12 +601,12 @@ func (v *VoiceConnection) udpOpen() (err error) {
 	rb := make([]byte, 74)
 	rlen, _, err := v.udpConn.ReadFromUDP(rb)
 	if err != nil {
-		v.log(LogWarning, "udp read error, %s, %s", addr.String(), err)
+		v.LogError("udp read error, %s, %s", addr.String(), err)
 		return
 	}
 
 	if rlen < 74 {
-		v.log(LogWarning, "received udp packet too small")
+		v.LogWarn("received udp packet too small")
 		return fmt.Errorf("received udp packet too small")
 	}
 
@@ -643,7 +630,7 @@ func (v *VoiceConnection) udpOpen() (err error) {
 	err = v.wsConn.WriteJSON(data)
 	v.wsMutex.Unlock()
 	if err != nil {
-		v.log(LogWarning, "udp write error, %#v, %s", data, err)
+		v.LogError("udp write error, %#v, %s", data, err)
 		return
 	}
 
@@ -657,7 +644,6 @@ func (v *VoiceConnection) udpOpen() (err error) {
 // udpKeepAlive sends a udp packet to keep the udp connection open
 // This is still a bit of a "proof of concept"
 func (v *VoiceConnection) udpKeepAlive(udpConn *net.UDPConn, close <-chan struct{}, i time.Duration) {
-
 	if udpConn == nil || close == nil {
 		return
 	}
@@ -676,7 +662,7 @@ func (v *VoiceConnection) udpKeepAlive(udpConn *net.UDPConn, close <-chan struct
 
 		_, err = udpConn.Write(packet)
 		if err != nil {
-			v.log(LogError, "write error, %s", err)
+			v.LogError("write error, %s", err)
 			return
 		}
 
@@ -692,7 +678,6 @@ func (v *VoiceConnection) udpKeepAlive(udpConn *net.UDPConn, close <-chan struct
 // opusSender will listen on the given channel and send any
 // pre-encoded opus audio to Discord.  Supposedly.
 func (v *VoiceConnection) opusSender(udpConn *net.UDPConn, close <-chan struct{}, opus <-chan []byte, rate, size int) {
-
 	if udpConn == nil || close == nil {
 		return
 	}
@@ -742,7 +727,7 @@ func (v *VoiceConnection) opusSender(udpConn *net.UDPConn, close <-chan struct{}
 		if !speaking {
 			err := v.Speaking(true)
 			if err != nil {
-				v.log(LogError, "error sending speaking packet, %s", err)
+				v.LogError("error sending speaking packet, %s", err)
 			}
 		}
 
@@ -767,8 +752,8 @@ func (v *VoiceConnection) opusSender(udpConn *net.UDPConn, close <-chan struct{}
 		_, err := udpConn.Write(sendbuf)
 
 		if err != nil {
-			v.log(LogError, "udp write error, %s", err)
-			v.log(LogDebug, "voice struct: %#v\n", v)
+			v.LogError("udp write error, %s", err)
+			v.LogDebug("voice struct: %#v\n", v)
 			return
 		}
 
@@ -800,7 +785,6 @@ type Packet struct {
 // and sends them across the given channel
 // NOTE :: This function may change names later.
 func (v *VoiceConnection) opusReceiver(udpConn *net.UDPConn, close <-chan struct{}, c chan *Packet) {
-
 	if udpConn == nil || close == nil {
 		return
 	}
@@ -819,8 +803,8 @@ func (v *VoiceConnection) opusReceiver(udpConn *net.UDPConn, close <-chan struct
 			v.RUnlock()
 			if sameConnection {
 
-				v.log(LogError, "udp read error, %s, %s", v.endpoint, err)
-				v.log(LogDebug, "voice struct: %#v\n", v)
+				v.LogError("udp read error, %s, %s", v.endpoint, err)
+				v.LogDebug("voice struct: %#v\n", v)
 
 				go v.reconnect()
 			}
@@ -881,12 +865,11 @@ func (v *VoiceConnection) opusReceiver(udpConn *net.UDPConn, close <-chan struct
 // It will be cleaned up once a proven stable option is flushed out.
 // aka: this is ugly shit code, please don't judge too harshly.
 func (v *VoiceConnection) reconnect() {
-
-	v.log(LogInformational, "called")
+	v.LogDebug("called")
 
 	v.Lock()
 	if v.reconnecting {
-		v.log(LogInformational, "already reconnecting to channel %s, exiting", v.ChannelID)
+		v.LogWarn("already reconnecting to channel %s, exiting", v.ChannelID)
 		v.Unlock()
 		return
 	}
@@ -904,7 +887,6 @@ func (v *VoiceConnection) reconnect() {
 
 	wait := time.Duration(1)
 	for {
-
 		<-time.After(wait * time.Second)
 		wait *= 2
 		if wait > 600 {
@@ -912,19 +894,19 @@ func (v *VoiceConnection) reconnect() {
 		}
 
 		if v.session.DataReady == false || v.session.wsConn == nil {
-			v.log(LogInformational, "cannot reconnect to channel %s with unready session", v.ChannelID)
+			v.LogWarn("cannot reconnect to channel %s with unready session", v.ChannelID)
 			continue
 		}
 
-		v.log(LogInformational, "trying to reconnect to channel %s", v.ChannelID)
+		v.LogInfo("trying to reconnect to channel %s", v.ChannelID)
 
 		_, err := v.session.ChannelVoiceJoin(v.GuildID, v.ChannelID, v.mute, v.deaf)
 		if err == nil {
-			v.log(LogInformational, "successfully reconnected to channel %s", v.ChannelID)
+			v.LogInfo("successfully reconnected to channel %s", v.ChannelID)
 			return
 		}
 
-		v.log(LogInformational, "error reconnecting to channel %s, %s", v.ChannelID, err)
+		v.LogError("error reconnecting to channel %s, %s", v.ChannelID, err)
 
 		// if the reconnect above didn't work lets just send a disconnect
 		// packet to reset things.
@@ -934,8 +916,7 @@ func (v *VoiceConnection) reconnect() {
 		err = v.session.wsConn.WriteJSON(data)
 		v.session.wsMutex.Unlock()
 		if err != nil {
-			v.log(LogError, "error sending disconnect packet, %s", err)
+			v.LogError("error sending disconnect packet, %s", err)
 		}
-
 	}
 }
